Finalize crowdsale functionality
