+++ src/App.js
@@ -21,7 +21,7 @@ class App extends Component {
           <Route exact path="/" component={crowdsaleAddr?Crowdsale:Home}/>
           <Route exact path="/crowdsale" component={Crowdsale}/>
           <Route exact path="/invest" component={Invest}/>
-          <Route exact path="/manage" component={Manage}/>
+          <Route exact path="/manage/:addr" component={Manage}/>
           <Route path="/1" component={stepOne}/>
           <Route path="/2" component={stepTwo}/>
           <Route path="/3" component={stepThree}/>
