+++ src/components/index.js
@@ -1,10 +1,10 @@
 export { Footer } from './Footer';
 export { Header } from './Header';
 export { Home } from './Home';
-export { Manage } from './manage'
 export { stepOne } from './stepOne'
 export { stepTwo } from './stepTwo'
 export { stepThree } from './stepThree'
 export { stepFour } from './stepFour'
 export { Crowdsale } from './crowdsale'
 export { Invest } from './invest'
+export { Manage } from './manage'
